Implemented banning/kicking users off a stream Closes #83

+++ backend/src/services/stream/removeUserFromStreams.ts
@@ -3,13 +3,23 @@ import { BroadcastService, FeedCacheService } from "..";
 
 export const removeUserFromStreams = async (user: string): Promise<void> => {
   const stream = await ParticipantDAL.getCurrentStreamFor(user);
+  const participant = await ParticipantDAL.getById(user);
 
   try {
-    if (stream) {
-      BroadcastService.broadcastParticipantLeft(user, stream);
+    if (participant) {
+      await ParticipantDAL.updateOne(user, {
+        hasLeftStream: true,
+      });
+
+      if (stream) {
+        BroadcastService.broadcastParticipantLeft(user, stream);
+
+        if (participant?.role === "streamer") {
+          ParticipantDAL.deleteParticipant(user);
+        }
+      }
     }
 
-    await ParticipantDAL.deleteParticipant(user);
     await FeedCacheService.removeServedStreams(user);
   } catch (e) {
     console.error(e);
